% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{dependencies}
\alias{dependencies}
\alias{install_pyoscar}
\title{Install Dependencies}
\usage{
install_pyoscar(envname = NULL, ...)
}
\arguments{
\item{envname}{The name, or full path, of the environment in which Python packages are to be installed. When NULL (the default), the active environment as set by the RETICULATE_PYTHON_ENV variable will be used; if that is unset, then the r-reticulate environment will be used.}

\item{...}{Additional arguments passed to conda_install() or virtualenv_install().}
}
\description{
Install Dependencies
}
