% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_report.R
\name{get_station_report}
\alias{get_station_report}
\title{Get OSCAR station report summary by providing it's identifier}
\usage{
get_station_report(identifier = "0-20000-0-71758")
}
\arguments{
\item{identifier}{}
}
\value{
a \link[tibble:tibble-package]{tibble}. Summary report of the station with identifier
}
\description{
Get OSCAR station report summary by providing it's identifier
}
\examples{
library(oscarr)
get_station_report(identifier = "0-20000-0-71758")
}
