% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations.R
\name{get_stations}
\alias{get_stations}
\title{Title}
\usage{
get_stations(country = "BFA")
}
\arguments{
\item{country}{Official ISO 3C of country name}
}
\value{
A tibble. OSCAR Station list of the specified country
}
\description{
Title
}
