% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{oscar_station_type}
\alias{oscar_station_type}
\title{Get WMO OSCAR station type name}
\usage{
oscar_station_type()
}
\value{
A character vector of WMO OSCAR station type
}
\description{
Get WMO OSCAR station type name
}
\examples{
library(oscarr)
oscar_station_type()
oscar_station_type()[8] # landFixed 

}
