% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations.R
\name{get_stations}
\alias{get_stations}
\title{Retrieve WMO OSCAR station metadata}
\usage{
get_stations(
  program = NULL,
  country = "BFA",
  station_type = NULL,
  wigos_id = NULL
)
}
\arguments{
\item{program}{WMO OSCAR program/network. Example : "GBON","RBCN", etc.}

\item{country}{Official ISO 3166-1 alpha-3 code (3 letters)}

\item{station_type}{WMO OSCAR station type. Example : "landFixed","airMobile" etc. Use \code{oscarr::oscar_station_type()} for a complete list.}

\item{identifier}{WIGOS Station identifier. Example : "0-854-0-109".}
}
\value{
A tibble of metadata of all matching stations
}
\description{
Retrieve WMO OSCAR station metadata
}
\examples{
\dontrun{
library(oscarr)
country <- "NGA" # Nigeria ISO 3166-1 alpha-3 code
nga_stns <- get_stations(country = country)
View(nga_stns)
}


}
