% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscarr-package.R
\docType{package}
\name{oscarr-package}
\alias{oscarr}
\alias{oscarr-package}
\title{oscarr: R interface to pyoscar}
\description{
provides access to the WMO OSCAR HTTP API..
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oousmane/oscarr}
  \item Report bugs at \url{https://github.com/oousmane/oscarr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
